Move the kickstart specification to pyanaconda.core.kickstart
The kickstart specification doesn't depend on DBus and it should be
stable, so it can be part of the core package.
