+++ pyanaconda/core/kickstart.py
