+++ pyanaconda/modules/base.py
@@ -26,7 +26,7 @@ from pyanaconda.core.timer import Timer
 from pyanaconda.dbus import DBus
 from pyanaconda.task import publish_task
 from pyanaconda.core.signal import Signal
-from pyanaconda.modules.base_kickstart import NoKickstartSpecification, \
+from pyanaconda.core.kickstart import NoKickstartSpecification, \
     KickstartSpecificationHandler, KickstartSpecificationParser
 
 from pyanaconda.anaconda_loggers import get_module_logger
